split db related functions out
