+++ light-client-lib/Cargo.toml
@@ -1,81 +1,85 @@
 [package]
-name = "ckb-light-client-lib"
-version = "0.5.4"
 authors = ["Nervos Core Dev <dev@nervos.org>"]
-edition = "2021"
-license = "MIT"
 description = "A CKB light client based on FlyClient."
+edition = "2021"
 homepage = "https://github.com/nervosnetwork/ckb-light-client"
+license = "MIT"
+name = "ckb-light-client-lib"
 repository = "https://github.com/nervosnetwork/ckb-light-client"
+version = "0.5.4"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 ckb-app-config = "1"
-ckb-constant = "1"
-ckb-types = "1"
-ckb-network = "1"
-ckb-jsonrpc-types = "1"
-ckb-error = "1"
-ckb-script = "1"
 ckb-chain-spec = "1"
-ckb-traits = "1"
-ckb-resource = "1"
-ckb-verification = "1"
-ckb-hash = "1"
-ckb-systemtime = "1"
-ckb-dao-utils = "1"
+ckb-constant = "1"
 ckb-dao = "1"
+ckb-dao-utils = "1"
+ckb-error = "1"
+ckb-hash = "1"
+ckb-jsonrpc-types = "1"
+ckb-network = "1"
+ckb-resource = "1"
+ckb-script = "1"
+ckb-systemtime = "1"
+ckb-traits = "1"
+ckb-types = "1"
+ckb-verification = "1"
 
+ckb-merkle-mountain-range = "0.5.1"
+dashmap = "5.3"
+golomb-coded-set = "0.2.1"
+governor = "0.6.3"
+linked-hash-map = "0.5.6"
 log = "0.4.14"
+numext-fixed-uint = {version = "0.1", features = [
+  "support_rand",
+  "support_heapsize",
+  "support_serde",
+]}
 path-clean = "0.1.0"
 rand = "0.8.5"
-dashmap = "5.3"
-linked-hash-map = "0.5.6"
-governor = "0.6.3"
-serde = { version = "1.0.136", features = ["derive"] }
-ckb-merkle-mountain-range = "0.5.1"
-golomb-coded-set = "0.2.1"
-numext-fixed-uint = { version = "0.1", features = [
-    "support_rand",
-    "support_heapsize",
-    "support_serde",
-] }
+serde = {version = "1.0.136", features = ["derive"]}
 
 anyhow = "1.0.56"
 thiserror = "1.0.30"
+tokio = {version = "1.20"}
 toml = "0.5.8"
-tokio = { version = "1.20" }
+
+[features]
+default = ["rocksdb"]
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-rocksdb = { package = "ckb-rocksdb", version = "=0.21.1", features = [
-    "snappy",
-], default-features = false }
+rocksdb = {package = "ckb-rocksdb", version = "=0.21.1", features = [
+  "snappy",
+], default-features = false, optional = true}
+rusqlite = {version = "0.38.0", features = ["bundled"], optional = true}
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-web-time = "1.1.0"
-wasm-bindgen = "0.2.63"
+console_error_panic_hook = {version = "0.1.7"}
 idb = "0.6"
+light-client-db-common = {path = "../wasm/light-client-db-common"}
 serde-wasm-bindgen = "0.6.5"
-light-client-db-common = { path = "../wasm/light-client-db-common" }
-web-sys = "0.3.72"
-console_error_panic_hook = { version = "0.1.7" }
 serde_json = "1.0.134"
+wasm-bindgen = "0.2.63"
+web-sys = "0.3.72"
+web-time = "1.1.0"
 
 [target.'cfg(target_arch = "wasm32")'.dev-dependencies]
 wasm-bindgen-test = "0.3.45"
 
 [dev-dependencies]
-ckb-shared = "1"
 ckb-chain = "1"
-ckb-tx-pool = "1"
-ckb-store = "1"
-ckb-systemtime = { version = "1", features = ["enable_faketime"] }
-tempfile = "3.0"
-rand = "0.8"
-serde_json = "1.0"
-lazy_static = "1.3.0"
-env_logger = "0.11"
 ckb-db = "1"
 ckb-db-schema = "1"
+ckb-shared = "1"
+ckb-store = "1"
+ckb-systemtime = {version = "1", features = ["enable_faketime"]}
+ckb-tx-pool = "1"
+env_logger = "0.11"
+lazy_static = "1.3.0"
+rand = "0.8"
+serde_json = "1.0"
+tempfile = "3.0"
