+++ light-client-lib/src/storage/db/native_rocksdb.rs
@@ -1,0 +1,605 @@
+use super::super::{
+    extract_raw_data, parse_matched_blocks, BlockNumber, Byte32, CellIndex, CellType, CpIndex, Key,
+    KeyPrefix, MatchedBlock, MatchedBlocks, Script, FILTER_SCRIPTS_KEY, MATCHED_FILTER_BLOCKS_KEY,
+    MIN_FILTERED_BLOCK_NUMBER,
+};
+use crate::{error::Result, storage::{ScriptStatus, ScriptType, SetScriptsCommand, TxIndex}};
+use ckb_traits::{CellDataProvider, HeaderProvider};
+use ckb_types::{
+    bytes::Bytes,
+    core::{
+        cell::{CellMeta, CellProvider, CellStatus},
+        HeaderView, TransactionInfo,
+    },
+    packed::{self, CellOutput, Header, OutPoint},
+    prelude::*,
+};
+use rocksdb::{
+    ops::{Delete, GetPinned},
+    prelude::{Get, Iterate, Open, Put, WriteOps},
+    DBPinnableSlice, Direction, IteratorMode, Options, Snapshot, WriteBatch, DB,
+};
+use std::{path::Path, sync::Arc};
+
+pub struct PinnedSlice<'a> {
+    inner: rocksdb::DBPinnableSlice<'a>,
+}
+
+impl<'a> From<DBPinnableSlice<'a>> for PinnedSlice<'a> {
+    fn from(value: DBPinnableSlice<'a>) -> Self {
+        PinnedSlice { inner: value }
+    }
+}
+
+impl AsRef<[u8]> for PinnedSlice<'_> {
+    fn as_ref(&self) -> &[u8] {
+        self.inner.as_ref()
+    }
+}
+
+#[derive(Clone)]
+pub struct Storage {
+    pub(crate) db: Arc<DB>,
+}
+
+impl Storage {
+    pub fn new<P: AsRef<Path>>(path: P) -> Self {
+        let mut opts = Options::default();
+        opts.create_if_missing(true);
+        opts.set_max_total_wal_size(128 * 1024 * 1024);
+        opts.set_write_buffer_size(128 * 1024 * 1024);
+        opts.set_max_write_buffer_number(2);
+        let db = Arc::new(DB::open(&opts, path).expect("Failed to open rocksdb"));
+        Self { db }
+    }
+
+    pub fn batch(&self) -> Batch {
+        Batch {
+            db: Arc::clone(&self.db),
+            wb: WriteBatch::default(),
+        }
+    }
+
+    pub fn snapshot(&self) -> Snapshot<'_> {
+        self.db.snapshot()
+    }
+
+    pub fn put<K, V>(&self, key: K, value: V) -> Result<()>
+    where
+        K: AsRef<[u8]>,
+        V: AsRef<[u8]>,
+    {
+        self.db.put(key, value).map_err(Into::into)
+    }
+
+    pub fn get<K: AsRef<[u8]>>(&self, key: K) -> Result<Option<Vec<u8>>> {
+        self.db
+            .get(key.as_ref())
+            .map(|v| v.map(|vi| vi.to_vec()))
+            .map_err(Into::into)
+    }
+
+    pub fn get_pinned<K>(&self, key: K) -> Result<Option<PinnedSlice<'_>>>
+    where
+        K: AsRef<[u8]>,
+    {
+        self.db
+            .get_pinned(key)
+            .map_err(Into::into)
+            .map(|a| a.map(Into::into))
+    }
+
+    pub fn delete<K: AsRef<[u8]>>(&self, key: K) -> Result<()> {
+        self.db.delete(key).map_err(Into::into)
+    }
+
+    pub fn is_filter_scripts_empty(&self) -> bool {
+        let key_prefix = Key::Meta(FILTER_SCRIPTS_KEY).into_vec();
+        let mode = IteratorMode::From(key_prefix.as_ref(), Direction::Forward);
+        self.db
+            .iterator(mode)
+            .take_while(|(key, _value)| key.starts_with(&key_prefix))
+            .next()
+            .is_none()
+    }
+
+    pub fn get_filter_scripts(&self) -> Vec<ScriptStatus> {
+        let key_prefix = Key::Meta(FILTER_SCRIPTS_KEY).into_vec();
+        let mode = IteratorMode::From(key_prefix.as_ref(), Direction::Forward);
+
+        self.db
+            .iterator(mode)
+            .take_while(|(key, _value)| key.starts_with(&key_prefix))
+            .map(|(key, value)| {
+                let script = Script::from_slice(&key[key_prefix.len()..key.len() - 1])
+                    .expect("stored Script");
+                let script_type = match key[key.len() - 1] {
+                    0 => ScriptType::Lock,
+                    1 => ScriptType::Type,
+                    _ => panic!("invalid script type"),
+                };
+                let block_number = BlockNumber::from_be_bytes(
+                    value.as_ref().try_into().expect("stored BlockNumber"),
+                );
+                ScriptStatus {
+                    script,
+                    script_type,
+                    block_number,
+                }
+            })
+            .collect()
+    }
+
+    pub fn update_filter_scripts(&self, scripts: Vec<ScriptStatus>, command: SetScriptsCommand) {
+        let mut should_filter_genesis_block = false;
+        let mut batch = self.batch();
+        let key_prefix = Key::Meta(FILTER_SCRIPTS_KEY).into_vec();
+
+        match command {
+            SetScriptsCommand::All => {
+                should_filter_genesis_block = scripts.iter().any(|ss| ss.block_number == 0);
+                let mode = IteratorMode::From(key_prefix.as_ref(), Direction::Forward);
+
+                self.db
+                    .iterator(mode)
+                    .take_while(|(key, _value)| key.starts_with(&key_prefix))
+                    .for_each(|(key, _value)| {
+                        batch.delete(key).expect("batch delete should be ok");
+                    });
+
+                for ss in scripts {
+                    let key = [
+                        key_prefix.as_ref(),
+                        ss.script.as_slice(),
+                        match ss.script_type {
+                            ScriptType::Lock => &[0],
+                            ScriptType::Type => &[1],
+                        },
+                    ]
+                    .concat();
+                    batch
+                        .put(key, ss.block_number.to_be_bytes())
+                        .expect("batch put should be ok");
+                }
+            }
+            SetScriptsCommand::Partial => {
+                if scripts.is_empty() {
+                    return;
+                }
+                let min_script_block_number = scripts.iter().map(|ss| ss.block_number).min();
+                should_filter_genesis_block = min_script_block_number == Some(0);
+
+                for ss in scripts {
+                    let key = [
+                        key_prefix.as_ref(),
+                        ss.script.as_slice(),
+                        match ss.script_type {
+                            ScriptType::Lock => &[0],
+                            ScriptType::Type => &[1],
+                        },
+                    ]
+                    .concat();
+                    batch
+                        .put(key, ss.block_number.to_be_bytes())
+                        .expect("batch put should be ok");
+                }
+            }
+            SetScriptsCommand::Delete => {
+                if scripts.is_empty() {
+                    return;
+                }
+
+                for ss in scripts {
+                    let key = [
+                        key_prefix.as_ref(),
+                        ss.script.as_slice(),
+                        match ss.script_type {
+                            ScriptType::Lock => &[0],
+                            ScriptType::Type => &[1],
+                        },
+                    ]
+                    .concat();
+                    batch.delete(key).expect("batch delete should be ok");
+                }
+            }
+        }
+
+        batch.commit().expect("batch commit should be ok");
+
+        self.update_min_filtered_block_number_by_scripts();
+        self.clear_matched_blocks();
+
+        if should_filter_genesis_block {
+            let block = self.get_genesis_block();
+            self.filter_block(block);
+        }
+    }
+
+    fn update_min_filtered_block_number_by_scripts(&self) {
+        let key_prefix = Key::Meta(FILTER_SCRIPTS_KEY).into_vec();
+        let mode = IteratorMode::From(key_prefix.as_ref(), Direction::Forward);
+
+        let min_block_number = self
+            .db
+            .iterator(mode)
+            .take_while(|(key, _value)| key.starts_with(&key_prefix))
+            .map(|(_key, value)| {
+                BlockNumber::from_be_bytes(value.as_ref().try_into().expect("stored BlockNumber"))
+            })
+            .min();
+
+        if let Some(n) = min_block_number {
+            self.update_min_filtered_block_number(n);
+        }
+    }
+
+    // get scripts hash that should be filtered below the given block number
+    pub fn get_scripts_hash(&self, block_number: BlockNumber) -> Vec<Byte32> {
+        let key_prefix = Key::Meta(FILTER_SCRIPTS_KEY).into_vec();
+        let mode = IteratorMode::From(key_prefix.as_ref(), Direction::Forward);
+
+        self.db
+            .iterator(mode)
+            .take_while(|(key, _value)| key.starts_with(&key_prefix))
+            .filter_map(|(key, value)| {
+                let stored_block_number = BlockNumber::from_be_bytes(
+                    value.as_ref().try_into().expect("stored BlockNumber"),
+                );
+                if stored_block_number < block_number {
+                    let script = Script::from_slice(&key[key_prefix.len()..key.len() - 1])
+                        .expect("stored Script");
+                    Some(script.calc_script_hash())
+                } else {
+                    None
+                }
+            })
+            .collect()
+    }
+
+    fn clear_matched_blocks(&self) {
+        let key_prefix = Key::Meta(MATCHED_FILTER_BLOCKS_KEY).into_vec();
+        let mode = IteratorMode::From(key_prefix.as_ref(), Direction::Forward);
+        let mut batch = self.batch();
+        for (key, _) in self
+            .db
+            .iterator(mode)
+            .take_while(|(key, _value)| key.starts_with(&key_prefix))
+        {
+            batch.delete(key).expect("batch delete should be ok");
+        }
+        batch.commit().expect("batch commit should be ok");
+    }
+
+    fn get_matched_blocks(&self, direction: Direction) -> Option<MatchedBlocks> {
+        let key_prefix = Key::Meta(MATCHED_FILTER_BLOCKS_KEY).into_vec();
+        let iter_from = match direction {
+            Direction::Forward => key_prefix.clone(),
+            Direction::Reverse => {
+                let mut key = key_prefix.clone();
+                key.extend(u64::MAX.to_be_bytes());
+                key
+            }
+        };
+        let mode = IteratorMode::From(iter_from.as_ref(), direction);
+        self.db
+            .iterator(mode)
+            .take_while(|(key, _value)| key.starts_with(&key_prefix))
+            .map(|(key, value)| {
+                let mut u64_bytes = [0u8; 8];
+                u64_bytes.copy_from_slice(&key[key_prefix.len()..]);
+                let start_number = u64::from_be_bytes(u64_bytes);
+                let (blocks_count, raw_blocks) = parse_matched_blocks(&value);
+                let blocks = raw_blocks
+                    .into_iter()
+                    .map(|(hash, proved)| MatchedBlock { hash, proved })
+                    .collect();
+                MatchedBlocks {
+                    start_number,
+                    blocks_count,
+                    blocks,
+                }
+            })
+            .next()
+    }
+
+    pub fn get_earliest_matched_blocks(&self) -> Option<MatchedBlocks> {
+        self.get_matched_blocks(Direction::Forward)
+    }
+
+    pub fn get_latest_matched_blocks(&self) -> Option<MatchedBlocks> {
+        self.get_matched_blocks(Direction::Reverse)
+    }
+
+    pub fn get_check_points(&self, start_index: CpIndex, limit: usize) -> Vec<Byte32> {
+        let start_key = Key::CheckPointIndex(start_index).into_vec();
+        let key_prefix = [KeyPrefix::CheckPointIndex as u8];
+        let mode = IteratorMode::From(start_key.as_ref(), Direction::Forward);
+        self.db
+            .iterator(mode)
+            .take_while(|(key, _value)| key.starts_with(&key_prefix))
+            .take(limit)
+            .map(|(_key, value)| Byte32::from_slice(&value).expect("stored block filter hash"))
+            .collect()
+    }
+
+    pub fn update_block_number(&self, block_number: BlockNumber) {
+        let key_prefix = Key::Meta(FILTER_SCRIPTS_KEY).into_vec();
+        let mode = IteratorMode::From(key_prefix.as_ref(), Direction::Forward);
+
+        let mut batch = self.batch();
+        self.db
+            .iterator(mode)
+            .take_while(|(key, _value)| key.starts_with(&key_prefix))
+            .for_each(|(key, value)| {
+                let stored_block_number = BlockNumber::from_be_bytes(
+                    value.as_ref().try_into().expect("stored BlockNumber"),
+                );
+                if stored_block_number < block_number {
+                    batch
+                        .put(key, block_number.to_be_bytes())
+                        .expect("batch put should be ok")
+                }
+            });
+        batch.commit().expect("batch commit should be ok");
+    }
+
+    /// Rollback filtered block data to specified block number
+    ///
+    /// N.B. The specified block will be removed.
+    pub fn rollback_to_block(&self, to_number: BlockNumber) {
+        let scripts = self.get_filter_scripts();
+        let mut batch = self.batch();
+
+        for ss in scripts {
+            if ss.block_number >= to_number {
+                let script = ss.script;
+                let mut key_prefix = vec![match ss.script_type {
+                    ScriptType::Lock => KeyPrefix::TxLockScript as u8,
+                    ScriptType::Type => KeyPrefix::TxTypeScript as u8,
+                }];
+                key_prefix.extend_from_slice(&extract_raw_data(&script));
+                let mut start_key = key_prefix.clone();
+                start_key.extend_from_slice(BlockNumber::MAX.to_be_bytes().as_ref());
+                let mode = IteratorMode::From(start_key.as_ref(), Direction::Reverse);
+                let key_prefix_len = key_prefix.len();
+
+                self.db
+                    .iterator(mode)
+                    .take_while(|(key, _value)| {
+                        key.starts_with(&key_prefix)
+                            && BlockNumber::from_be_bytes(
+                                key[key_prefix_len..key_prefix_len + 8]
+                                    .try_into()
+                                    .expect("stored BlockNumber"),
+                            ) >= to_number
+                    })
+                    .for_each(|(key, value)| {
+                        let block_number = BlockNumber::from_be_bytes(
+                            key[key_prefix_len..key_prefix_len + 8]
+                                .try_into()
+                                .expect("stored BlockNumber"),
+                        );
+                        log::debug!("rollback {}", block_number);
+                        let tx_index = TxIndex::from_be_bytes(
+                            key[key_prefix_len + 8..key_prefix_len + 12]
+                                .try_into()
+                                .expect("stored TxIndex"),
+                        );
+                        let cell_index = CellIndex::from_be_bytes(
+                            key[key_prefix_len + 12..key_prefix_len + 16]
+                                .try_into()
+                                .expect("stored CellIndex"),
+                        );
+                        let tx_hash =
+                            packed::Byte32Reader::from_slice_should_be_ok(&value).to_entity();
+                        if key[key_prefix_len + 16] == 0 {
+                            let (_, _, tx) = self
+                                .get_transaction(&tx_hash)
+                                .expect("stored transaction history");
+                            let input = tx.raw().inputs().get(cell_index as usize).unwrap();
+                            if let Some((
+                                generated_by_block_number,
+                                generated_by_tx_index,
+                                _previous_tx,
+                            )) = self.get_transaction(&input.previous_output().tx_hash())
+                            {
+                                let key = match ss.script_type {
+                                    ScriptType::Lock => Key::CellLockScript(
+                                        &script,
+                                        generated_by_block_number,
+                                        generated_by_tx_index,
+                                        input.previous_output().index().unpack(),
+                                    ),
+                                    ScriptType::Type => Key::CellTypeScript(
+                                        &script,
+                                        generated_by_block_number,
+                                        generated_by_tx_index,
+                                        input.previous_output().index().unpack(),
+                                    ),
+                                };
+                                batch
+                                    .put_kv(key, input.previous_output().tx_hash().as_slice())
+                                    .expect("batch put should be ok");
+                            };
+                            // delete tx history
+                            let key = match ss.script_type {
+                                ScriptType::Lock => Key::TxLockScript(
+                                    &script,
+                                    block_number,
+                                    tx_index,
+                                    cell_index,
+                                    CellType::Input,
+                                ),
+                                ScriptType::Type => Key::TxTypeScript(
+                                    &script,
+                                    block_number,
+                                    tx_index,
+                                    cell_index,
+                                    CellType::Input,
+                                ),
+                            }
+                            .into_vec();
+                            batch.delete(key).expect("batch delete should be ok");
+                        } else {
+                            // delete utxo
+                            let key = match ss.script_type {
+                                ScriptType::Lock => {
+                                    Key::CellLockScript(&script, block_number, tx_index, cell_index)
+                                }
+                                ScriptType::Type => {
+                                    Key::CellTypeScript(&script, block_number, tx_index, cell_index)
+                                }
+                            }
+                            .into_vec();
+                            batch.delete(key).expect("batch delete should be ok");
+
+                            // delete tx history
+                            let key = match ss.script_type {
+                                ScriptType::Lock => Key::TxLockScript(
+                                    &script,
+                                    block_number,
+                                    tx_index,
+                                    cell_index,
+                                    CellType::Output,
+                                ),
+                                ScriptType::Type => Key::TxTypeScript(
+                                    &script,
+                                    block_number,
+                                    tx_index,
+                                    cell_index,
+                                    CellType::Output,
+                                ),
+                            }
+                            .into_vec();
+                            batch.delete(key).expect("batch delete should be ok");
+                        };
+                    });
+
+                // update script filter block number
+                {
+                    let mut key = Key::Meta(FILTER_SCRIPTS_KEY).into_vec();
+                    key.extend_from_slice(script.as_slice());
+                    key.extend_from_slice(match ss.script_type {
+                        ScriptType::Lock => &[0],
+                        ScriptType::Type => &[1],
+                    });
+                    let value = to_number.to_be_bytes().to_vec();
+                    batch.put(key, value).expect("batch put should be ok");
+                }
+            }
+        }
+
+        // we should also sync block filters again
+        if self.get_min_filtered_block_number() >= to_number {
+            batch
+                .put(
+                    Key::Meta(MIN_FILTERED_BLOCK_NUMBER).into_vec(),
+                    to_number.saturating_sub(1).to_le_bytes(),
+                )
+                .expect("batch put should be ok");
+        }
+
+        batch.commit().expect("batch commit should be ok");
+    }
+}
+
+pub struct Batch {
+    db: Arc<DB>,
+    wb: WriteBatch,
+}
+
+impl Batch {
+    pub fn put_kv<K: Into<Vec<u8>>, V: Into<Vec<u8>>>(&mut self, key: K, value: V) -> Result<()> {
+        self.put(Into::<Vec<u8>>::into(key), Into::<Vec<u8>>::into(value))
+    }
+
+    pub fn put<K: AsRef<[u8]>, V: AsRef<[u8]>>(&mut self, key: K, value: V) -> Result<()> {
+        self.wb.put(key, value)?;
+        Ok(())
+    }
+
+    pub fn delete<K: AsRef<[u8]>>(&mut self, key: K) -> Result<()> {
+        self.wb.delete(key.as_ref())?;
+        Ok(())
+    }
+
+    pub fn commit(self) -> Result<()> {
+        self.db.write(&self.wb)?;
+        Ok(())
+    }
+}
+
+impl CellProvider for Storage {
+    // assume all cells are live and load data eagerly
+    fn cell(&self, out_point: &OutPoint, _eager_load: bool) -> CellStatus {
+        if let Some((block_number, tx_index, tx)) = self.get_transaction(&out_point.tx_hash()) {
+            let block_hash = Byte32::from_slice(
+                &self
+                    .get(Key::BlockNumber(block_number).into_vec())
+                    .expect("db get should be ok")
+                    .expect("stored block number / hash mapping"),
+            )
+            .expect("stored block hash should be OK");
+
+            let header = Header::from_slice(
+                &self
+                    .get(Key::BlockHash(&block_hash).into_vec())
+                    .expect("db get should be ok")
+                    .expect("stored block hash / header mapping")[..Header::TOTAL_SIZE],
+            )
+            .expect("stored header should be OK")
+            .into_view();
+
+            let output_index = out_point.index().unpack();
+            let tx = tx.into_view();
+            if let Some(cell_output) = tx.outputs().get(output_index) {
+                let output_data = tx
+                    .outputs_data()
+                    .get(output_index)
+                    .expect("output_data's index should be same as output")
+                    .raw_data();
+                let output_data_data_hash = CellOutput::calc_data_hash(&output_data);
+                let cell_meta = CellMeta {
+                    out_point: out_point.clone(),
+                    cell_output,
+                    transaction_info: Some(TransactionInfo {
+                        block_hash,
+                        block_epoch: header.epoch(),
+                        block_number,
+                        index: tx_index as usize,
+                    }),
+                    data_bytes: output_data.len() as u64,
+                    mem_cell_data: Some(output_data),
+                    mem_cell_data_hash: Some(output_data_data_hash),
+                };
+                return CellStatus::Live(cell_meta);
+            }
+        }
+        CellStatus::Unknown
+    }
+}
+
+impl CellDataProvider for Storage {
+    // we load all cells data eagerly in Storage's CellProivder impl
+    fn get_cell_data(&self, _out_point: &OutPoint) -> Option<Bytes> {
+        unreachable!()
+    }
+
+    fn get_cell_data_hash(&self, _out_point: &OutPoint) -> Option<Byte32> {
+        unreachable!()
+    }
+}
+
+impl HeaderProvider for Storage {
+    fn get_header(&self, hash: &Byte32) -> Option<HeaderView> {
+        self.get(Key::BlockHash(hash).into_vec())
+            .map(|v| {
+                v.map(|v| {
+                    Header::from_slice(&v[..Header::TOTAL_SIZE])
+                        .expect("stored Header")
+                        .into_view()
+                })
+            })
+            .expect("db get should be ok")
+    }
+}
